requirements.txt
torch
torchvision
torchaudio
numpy
opencv-python
Pillow
matplotlib
tqdm
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git